# api-specification.yaml
openapi: 3.0.3
info:
  title: CinemaAbyss API
  description: |
    API спецификация для системы CinemaAbyss, включающая монолит и микросервисы.
    
    Система CinemaAbyss представляет собой платформу для управления фильмами, пользователями, платежами и подписками.
    Архитектура системы включает монолитное приложение и выделенные микросервисы.
  version: 1.0.0
  contact:
    name: CinemaAbyss Team
    email: support@cinemaabyss.com
servers:
  - url: http://localhost:3200
    description: API Gateway (Proxy Service)
  - url: http://localhost:3280
    description: Монолитное приложение
  - url: http://localhost:3281
    description: Микросервис фильмов
  - url: http://localhost:3282
    description: Микросервис событий
...
paths:
...
  /api/movies:
    get:
      tags:
        - movies
      summary: Получение списка всех фильмов
      description: Возвращает список всех фильмов с их метаданными
      operationId: getAllMovies
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - movies
      summary: Создание нового фильма
      description: Добавляет новый фильм в систему
      operationId: createMovie
      requestBody:
        description: Данные фильма для создания
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '201':
          description: Фильм успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/users:
    get:
      tags:
        - users
      summary: Получение списка всех пользователей
      description: Возвращает список всех пользователей системы
      operationId: getAllUsers
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - users
      summary: Создание нового пользователя
      description: Регистрирует нового пользователя в системе
      operationId: createUser
      requestBody:
        description: Данные пользователя для создания
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Монолит - платежи
  /api/payments:
    get:
      tags:
        - payments
      summary: Получение списка всех платежей
      description: Возвращает список всех платежей в системе
      operationId: getAllPayments
      parameters:
        - name: user_id
          in: query
          description: ID пользователя для фильтрации платежей
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - payments
      summary: Создание нового платежа
      description: Регистрирует новый платеж в системе
      operationId: createPayment
      requestBody:
        description: Данные платежа для создания
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInput'
      responses:
        '201':
          description: Платеж успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Монолит - подписки
  /api/subscriptions:
    get:
      tags:
        - subscriptions
      summary: Получение списка всех подписок
      description: Возвращает список всех подписок в системе
      operationId: getAllSubscriptions
      parameters:
        - name: user_id
          in: query
          description: ID пользователя для фильтрации подписок
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - subscriptions
      summary: Создание новой подписки
      description: Регистрирует новую подписку в системе
      operationId: createSubscription
      requestBody:
        description: Данные подписки для создания
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionInput'
      responses:
        '201':
          description: Подписка успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Микросервис событий
...
components:
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор фильма
          example: 1
        title:
          type: string
          description: Название фильма
          example: "Inception"
        description:
          type: string
          description: Описание фильма
          example: "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O."
        genres:
          type: array
          description: Жанры фильма
          items:
            type: string
          example: ["Sci-Fi", "Action", "Thriller"]
        rating:
          type: number
          format: float
          description: Рейтинг фильма
          example: 8.8
      required:
        - id
        - title
        - description
        - rating
    MovieInput:
      type: object
      properties:
        title:
          type: string
          description: Название фильма
          example: "Inception"
        description:
          type: string
          description: Описание фильма
          example: "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O."
        genres:
          type: array
          description: Жанры фильма
          items:
            type: string
          example: ["Sci-Fi", "Action", "Thriller"]
        rating:
          type: number
          format: float
          description: Рейтинг фильма
          example: 8.8
      required:
        - title
        - description
        - rating
    User:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: "john_doe"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "john.doe@example.com"
      required:
        - id
        - username
        - email
    UserInput:
      type: object
      properties:
        username:
          type: string
          description: Имя пользователя
          example: "john_doe"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "john.doe@example.com"
      required:
        - username
        - email
    Payment:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор платежа
          example: 1
        user_id:
          type: integer
          description: Идентификатор пользователя, совершившего платеж
          example: 1
        amount:
          type: number
          format: float
          description: Сумма платежа
          example: 9.99
        timestamp:
          type: string
          format: date-time
          description: Время совершения платежа
          example: "2023-01-15T14:30:00Z"
      required:
        - id
        - user_id
        - amount
        - timestamp
    PaymentInput:
      type: object
      properties:
        user_id:
          type: integer
          description: Идентификатор пользователя, совершающего платеж
          example: 1
        amount:
          type: number
          format: float
          description: Сумма платежа
          example: 9.99
      required:
        - user_id
        - amount
    Subscription:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор подписки
          example: 1
        user_id:
          type: integer
          description: Идентификатор пользователя, оформившего подписку
          example: 1
        plan_type:
          type: string
          description: Тип плана подписки
          example: "premium"
        start_date:
          type: string
          format: date-time
          description: Дата начала подписки
          example: "2023-01-01T00:00:00Z"
        end_date:
          type: string
          format: date-time
          description: Дата окончания подписки
          example: "2023-12-31T23:59:59Z"
      required:
        - id
        - user_id
        - plan_type
        - start_date
        - end_date
    SubscriptionInput:
      type: object
      properties:
        user_id:
          type: integer
          description: Идентификатор пользователя, оформляющего подписку
          example: 1
        plan_type:
          type: string
          description: Тип плана подписки
          example: "premium"
        start_date:
          type: string
          format: date-time
          description: Дата начала подписки
          example: "2023-01-01T00:00:00Z"
        end_date:
          type: string
          format: date-time
          description: Дата окончания подписки
          example: "2023-12-31T23:59:59Z"
      required:
        - user_id
        - plan_type
        - start_date
        - end_date

    # Модели событий
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
          example: "Internal Server Error"
      required:
        - error